2_1_Gemini_API_Quickstart/examples/requirements.txt
google-genai>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
Pillow>=10.0.0
